refactor(summary): collapse newtype element modules into their wrappers
Movement and FailureRegression are each the element of a single newtype wrapper
(TopMovers, FailureRegressions) and are named nowhere else, so they collapse into
their wrapper module instead of standing as separate one-line files.
