--- solx-benchmark-converter/src/output/summary/movement.rs
@@ -1,13 +1,0 @@
-//!
-//! One row's movement between the main and PR toolchains.
-//!
-
-///
-/// One row's movement between the main and PR toolchains.
-///
-pub struct Movement {
-    pub label: String,
-    pub mode: String,
-    pub main: u64,
-    pub pr: u64,
-}
